@@ -14,12 +14,12 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  Token _paymentToken;
-  PaymentMethod _paymentMethod;
-  String _error;
-  final String _currentSecret = null; //set this yourself, e.g using curl
-  PaymentIntentResult _paymentIntent;
-  Source _source;
+  Token? _paymentToken;
+  PaymentMethod? _paymentMethod;
+  String? _error;
+  final String? _currentSecret = null; //set this yourself, e.g using curl
+  PaymentIntentResult? _paymentIntent;
+  Source? _source;
 
   ScrollController _controller = ScrollController();
 
@@ -29,7 +29,7 @@ class _MyAppState extends State<MyApp> {
     expYear: 21,
   );
 
-  GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey();
+  final GlobalKey<ScaffoldMessengerState> _scaffoldMessengerKey = GlobalKey<ScaffoldMessengerState>();
 
   @override
   initState() {
@@ -40,7 +40,7 @@ class _MyAppState extends State<MyApp> {
   }
 
   void setError(dynamic error) {
-    _scaffoldKey.currentState.showSnackBar(SnackBar(content: Text(error.toString())));
+    _scaffoldMessengerKey.currentState!.showSnackBar(SnackBar(content: Text(error.toString())));
     setState(() {
       _error = error.toString();
     });
@@ -49,8 +49,8 @@ class _MyAppState extends State<MyApp> {
   @override
   Widget build(BuildContext context) {
     return new MaterialApp(
-      home: new Scaffold(
-        key: _scaffoldKey,
+      scaffoldMessengerKey: _scaffoldMessengerKey,
+      home: Scaffold(
         appBar: new AppBar(
           title: new Text('Plugin example app'),
           actions: <Widget>[
@@ -71,7 +71,7 @@ class _MyAppState extends State<MyApp> {
           controller: _controller,
           padding: const EdgeInsets.all(20),
           children: <Widget>[
-            RaisedButton(
+            ElevatedButton(
               child: Text("Create Source"),
               onPressed: () {
                 StripePayment.createSourceWithParams(SourceParams(
@@ -80,40 +80,46 @@ class _MyAppState extends State<MyApp> {
                   currency: 'eur',
                   returnURL: 'example://stripe-redirect',
                 )).then((source) {
-                  _scaffoldKey.currentState.showSnackBar(SnackBar(content: Text('Received ${source.sourceId}')));
+                  _scaffoldMessengerKey.currentState!.showSnackBar(SnackBar(content: Text('Received ${source.sourceId}')));
                   setState(() {
                     _source = source;
                   });
-                }).catchError(setError);
+                }).catchError((error){
+                  setError(error);
+                });
               },
             ),
             Divider(),
-            RaisedButton(
+            ElevatedButton(
               child: Text("Create Token with Card Form"),
               onPressed: () {
                 StripePayment.paymentRequestWithCardForm(CardFormPaymentRequest()).then((paymentMethod) {
-                  _scaffoldKey.currentState.showSnackBar(SnackBar(content: Text('Received ${paymentMethod.id}')));
+                  _scaffoldMessengerKey.currentState!.showSnackBar(SnackBar(content: Text('Received ${paymentMethod.id}')));
                   setState(() {
                     _paymentMethod = paymentMethod;
                   });
-                }).catchError(setError);
+                }).catchError((error){
+                  setError(error);
+                });
               },
             ),
-            RaisedButton(
+            ElevatedButton(
               child: Text("Create Token with Card"),
               onPressed: () {
                 StripePayment.createTokenWithCard(
                   testCard,
                 ).then((token) {
-                  _scaffoldKey.currentState.showSnackBar(SnackBar(content: Text('Received ${token.tokenId}')));
+                  _scaffoldMessengerKey.currentState!.showSnackBar(SnackBar(content: Text('Received ${token.tokenId}')));
                   setState(() {
                     _paymentToken = token;
                   });
-                }).catchError(setError);
+                }).catchError((error){
+                  setError(error);
+                });
               },
             ),
             Divider(),
-            RaisedButton(
+            ElevatedButton(
               child: Text("Create Payment Method with Card"),
               onPressed: () {
                 StripePayment.createPaymentMethod(
@@ -121,14 +127,16 @@ class _MyAppState extends State<MyApp> {
                     card: testCard,
                   ),
                 ).then((paymentMethod) {
-                  _scaffoldKey.currentState.showSnackBar(SnackBar(content: Text('Received ${paymentMethod.id}')));
+                  _scaffoldMessengerKey.currentState!.showSnackBar(SnackBar(content: Text('Received ${paymentMethod.id}')));
                   setState(() {
                     _paymentMethod = paymentMethod;
                   });
-                }).catchError(setError);
+                }).catchError((error){
+                  setError(error);
+                });
               },
             ),
-            RaisedButton(
+            ElevatedButton(
               child: Text("Create Payment Method with existing token"),
               onPressed: _paymentToken == null
                   ? null
@@ -136,19 +144,21 @@ class _MyAppState extends State<MyApp> {
                       StripePayment.createPaymentMethod(
                         PaymentMethodRequest(
                           card: CreditCard(
-                            token: _paymentToken.tokenId,
+                            token: _paymentToken!.tokenId,
                           ),
                         ),
                       ).then((paymentMethod) {
-                        _scaffoldKey.currentState.showSnackBar(SnackBar(content: Text('Received ${paymentMethod.id}')));
+                        _scaffoldMessengerKey.currentState!.showSnackBar(SnackBar(content: Text('Received ${paymentMethod.id}')));
                         setState(() {
                           _paymentMethod = paymentMethod;
                         });
-                      }).catchError(setError);
+                      }).catchError((error){
+                        setError(error);
+                      });
                     },
             ),
             Divider(),
-            RaisedButton(
+            ElevatedButton(
               child: Text("Confirm Payment Intent"),
               onPressed: _paymentMethod == null || _currentSecret == null
                   ? null
@@ -156,33 +166,35 @@ class _MyAppState extends State<MyApp> {
                       StripePayment.confirmPaymentIntent(
                         PaymentIntent(
                           clientSecret: _currentSecret,
-                          paymentMethodId: _paymentMethod.id,
+                          paymentMethodId: _paymentMethod!.id,
                         ),
                       ).then((paymentIntent) {
-                        _scaffoldKey.currentState
-                            .showSnackBar(SnackBar(content: Text('Received ${paymentIntent.paymentIntentId}')));
+                        _scaffoldMessengerKey.currentState!.showSnackBar(SnackBar(content: Text('Received ${paymentIntent.paymentIntentId}')));
                         setState(() {
                           _paymentIntent = paymentIntent;
                         });
-                      }).catchError(setError);
+                      }).catchError((error){
+                        setError(error);
+                      });
                     },
             ),
-            RaisedButton(
+            ElevatedButton(
               child: Text("Authenticate Payment Intent"),
               onPressed: _currentSecret == null
                   ? null
                   : () {
-                      StripePayment.authenticatePaymentIntent(clientSecret: _currentSecret).then((paymentIntent) {
-                        _scaffoldKey.currentState
-                            .showSnackBar(SnackBar(content: Text('Received ${paymentIntent.paymentIntentId}')));
+                      StripePayment.authenticatePaymentIntent(clientSecret: _currentSecret!).then((paymentIntent) {
+                        _scaffoldMessengerKey.currentState!.showSnackBar(SnackBar(content: Text('Received ${paymentIntent.paymentIntentId}')));
                         setState(() {
                           _paymentIntent = paymentIntent;
                         });
-                      }).catchError(setError);
+                      }).catchError((error){
+                        setError(error);
+                      });
                     },
             ),
             Divider(),
-            RaisedButton(
+            ElevatedButton(
               child: Text("Native payment"),
               onPressed: () {
                 if (Platform.isIOS) {
@@ -205,18 +217,22 @@ class _MyAppState extends State<MyApp> {
                   ),
                 ).then((token) {
                   setState(() {
-                    _scaffoldKey.currentState.showSnackBar(SnackBar(content: Text('Received ${token.tokenId}')));
+                    _scaffoldMessengerKey.currentState!.showSnackBar(SnackBar(content: Text('Received ${token.tokenId}')));
                     _paymentToken = token;
                   });
-                }).catchError(setError);
+                }).catchError((error){
+                  setError(error);
+                });
               },
             ),
-            RaisedButton(
+            ElevatedButton(
               child: Text("Complete Native Payment"),
               onPressed: () {
                 StripePayment.completeNativePayRequest().then((_) {
-                  _scaffoldKey.currentState.showSnackBar(SnackBar(content: Text('Completed successfully')));
-                }).catchError(setError);
+                  _scaffoldMessengerKey.currentState!.showSnackBar(SnackBar(content: Text('Completed successfully')));
+                }).catchError((error){
+                  setError(error);
+                });
               },
             ),
             Divider(),
