@@ -7,13 +7,13 @@ enum RequiredShippingAddressFields { all, name, email, phone, postal_address }
 enum ShippingType { shipping, delivery, store_pickup, service_pickup }
 
 class ApplePayPaymentOptions {
-  List<RequiredBillingAddressFields> requiredBillingAddressFields;
-  List<RequiredShippingAddressFields> requiredShippingAddressFields;
-  List<ShippingMethod> shippingMethod;
-  String currencyCode;
-  String countryCode;
-  ShippingType shippingType;
-  List<ApplePayItem> items;
+  List<RequiredBillingAddressFields>? requiredBillingAddressFields;
+  List<RequiredShippingAddressFields>? requiredShippingAddressFields;
+  List<ShippingMethod>? shippingMethod;
+  String? currencyCode;
+  String? countryCode;
+  ShippingType? shippingType;
+  List<ApplePayItem>? items;
 
   ApplePayPaymentOptions(
       {this.requiredBillingAddressFields,
@@ -27,13 +27,13 @@ class ApplePayPaymentOptions {
   Map<dynamic, dynamic> get json {
     final Map<String, dynamic> data = new Map<String, dynamic>();
     if (this.requiredBillingAddressFields != null) {
-      data['requiredBillingAddressFields'] = this.requiredBillingAddressFields.map((b) => describeEnum(b)).toList();
+      data['requiredBillingAddressFields'] = this.requiredBillingAddressFields!.map((b) => describeEnum(b)).toList();
     }
     if (this.requiredShippingAddressFields != null) {
-      data['requiredShippingAddressFields'] = this.requiredShippingAddressFields.map((s) => describeEnum(s)).toList();
+      data['requiredShippingAddressFields'] = this.requiredShippingAddressFields!.map((s) => describeEnum(s)).toList();
     }
     if (this.shippingMethod != null) {
-      data['shippingMethod'] = this.shippingMethod.map((s) => s.toJson()).toList();
+      data['shippingMethod'] = this.shippingMethod!.map((s) => s.toJson()).toList();
     }
     data['currencyCode'] = currencyCode;
     if (this.shippingType != null) {
@@ -45,10 +45,10 @@ class ApplePayPaymentOptions {
 }
 
 class ShippingMethod {
-  String amount;
-  String detail;
-  String id;
-  String label;
+  String? amount;
+  String? detail;
+  String? id;
+  String? label;
 
   ShippingMethod({this.amount, this.detail, this.id, this.label});
 
@@ -72,9 +72,9 @@ class ShippingMethod {
 }
 
 class ApplePayItem {
-  String label;
-  String amount;
-  String type;
+  String? label;
+  String? amount;
+  String? type;
   ApplePayItem({this.label, this.amount, this.type});
 
   Map<String, dynamic> get json {
