@@ -1,10 +1,8 @@
 import 'dart:io';
 import 'package:flutter/foundation.dart';
 
-import 'package:flutter/material.dart';
-
 class Errors {
-  static Map<String, Map<String, String>> mapping = {
+  static Map<String, Map<String, String?>> mapping = {
     "busy": ErrorCode(
       errorCode: 'busy',
       description: 'Previous request is not completed',
@@ -123,9 +121,9 @@ class Errors {
 
 class ErrorCode {
   final String errorCode;
-  final String description;
+  final String? description;
 
-  ErrorCode({@required this.errorCode, this.description});
+  ErrorCode({required this.errorCode, this.description});
 
-  Map<String, String> toJson() => {"errorCode": errorCode, "description": description};
+  Map<String, String?> toJson() => {"errorCode": errorCode, "description": description};
 }
