@@ -1,23 +1,22 @@
-import 'package:flutter/material.dart';
 import 'package:stripe_payment/src/payment_method.dart';
 
 class PaymentIntent {
-  PaymentMethodRequest paymentMethod;
-  String paymentMethodId;
-  String returnURL;
-  String clientSecret;
+  PaymentMethodRequest? paymentMethod;
+  String? paymentMethodId;
+  String? returnURL;
+  String? clientSecret;
 
   PaymentIntent({
     this.paymentMethod,
     this.paymentMethodId,
     this.returnURL,
-    @required this.clientSecret,
+    this.clientSecret,
   });
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
     if (this.paymentMethod != null) {
-      data['paymentMethod'] = this.paymentMethod.toJson();
+      data['paymentMethod'] = this.paymentMethod!.toJson();
     }
     if (this.paymentMethodId != null) data['paymentMethodId'] = this.paymentMethodId;
     if (this.returnURL != null) data['returnURL'] = this.returnURL;
@@ -27,9 +26,9 @@ class PaymentIntent {
 }
 
 class PaymentIntentResult {
-  String status;
-  String paymentIntentId;
-  String paymentMethodId;
+  String? status;
+  String? paymentIntentId;
+  String? paymentMethodId;
 
   PaymentIntentResult({this.status, this.paymentIntentId, this.paymentMethodId});
 
@@ -51,9 +50,9 @@ class PaymentIntentResult {
 }
 
 class SetupIntentResult {
-  String status;
-  String paymentIntentId;
-  String setupIntentId;
+  String? status;
+  String? paymentIntentId;
+  String? setupIntentId;
 
   SetupIntentResult({
     this.status,
