@@ -1,6 +1,5 @@
 import 'dart:io';
 
-import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/foundation.dart';
 
@@ -48,22 +47,22 @@ class StripePayment {
       throw UnimplementedError();
     } else {
       if (Platform.isAndroid) {
-        return _channel.invokeMethod('canMakeAndroidPayPayments');
+        return _channel.invokeMethod('canMakeAndroidPayPayments') as Future<bool>;
       } else if (Platform.isIOS) {
         Map<String, dynamic> options = {"networks": networks};
-        return _channel.invokeMethod('canMakeApplePayPayments', options);
+        return _channel.invokeMethod('canMakeApplePayPayments', options) as Future<bool>;
       } else
         throw UnimplementedError();
     }
   }
 
-  static Future<bool> _deviceSupportsAndroidPay() => _channel.invokeMethod("deviceSupportsAndroidPay");
+  static Future<bool> _deviceSupportsAndroidPay() => _channel.invokeMethod("deviceSupportsAndroidPay") as Future<bool>;
 
-  static Future<bool> _deviceSupportsApplePay() => _channel.invokeMethod("deviceSupportsApplePay");
+  static Future<bool> _deviceSupportsApplePay() => _channel.invokeMethod("deviceSupportsApplePay") as Future<bool>;
 
   /// https://tipsi.github.io/tipsi-stripe/docs/paymentRequestWithNativePay.html
   static Future<Token> paymentRequestWithNativePay(
-      {@required AndroidPayPaymentRequest androidPayOptions, @required ApplePayPaymentOptions applePayOptions}) {
+      {required AndroidPayPaymentRequest androidPayOptions, required ApplePayPaymentOptions applePayOptions}) {
     if (kIsWeb) {
       throw UnimplementedError();
     } else {
@@ -83,7 +82,7 @@ class StripePayment {
 
   static Future<Token> _paymentRequestWithApplePay(ApplePayPaymentOptions options) async {
     final token = await _channel.invokeMethod("paymentRequestWithApplePay",
-        {"options": options.json, "items": options.items.map((item) => item.json).toList()});
+        {"options": options.json, "items": options.items!.map((item) => item.json).toList()});
     return Token.fromJson(token);
   }
 
@@ -146,8 +145,7 @@ class StripePayment {
   }
 
   /// https://tipsi.github.io/tipsi-stripe/docs/authenticatePaymentIntent.html
-  static Future<PaymentIntentResult> authenticatePaymentIntent({@required String clientSecret}) async {
-    assert(clientSecret != null);
+  static Future<PaymentIntentResult> authenticatePaymentIntent({required String clientSecret}) async {
     final result = await _channel.invokeMethod('authenticatePaymentIntent', {"clientSecret": clientSecret});
     return PaymentIntentResult.fromJson(result);
   }
@@ -161,8 +159,7 @@ class StripePayment {
   }
 
   /// https://tipsi.github.io/tipsi-stripe/docs/authenticateSetupIntent.html
-  static Future<SetupIntentResult> authenticateSetupIntent({@required String clientSecret}) async {
-    assert(clientSecret != null);
+  static Future<SetupIntentResult> authenticateSetupIntent({required String clientSecret}) async {
     final result = await _channel.invokeMethod('authenticateSetupIntent', {"clientSecret": clientSecret});
     return SetupIntentResult.fromJson(result);
   }
@@ -177,11 +174,11 @@ class StripePayment {
 }
 
 class StripeOptions {
-  final String publishableKey;
-  final String merchantId;
-  final String androidPayMode;
+  final String? publishableKey;
+  final String? merchantId;
+  final String? androidPayMode;
 
-  StripeOptions({@required this.publishableKey, this.merchantId, this.androidPayMode});
+  StripeOptions({required this.publishableKey, this.merchantId, this.androidPayMode});
 
   factory StripeOptions.fromJson(Map<String, dynamic> json) {
     return StripeOptions(
